See merge request s07-blockchain-nft-sub2/S07P22A607!17

+++ backend/src/main/java/com/trippiece/backend/api/service/TripService.java
@@ -8,17 +8,16 @@ import com.trippiece.backend.api.domain.entity.User;
 import com.trippiece.backend.api.domain.repository.RegionRepository;
 import com.trippiece.backend.api.domain.repository.TripRepository;
 import com.trippiece.backend.api.domain.repository.UserRepository;
+import com.trippiece.backend.exception.CustomException;
+import com.trippiece.backend.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.Duration;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -30,9 +29,9 @@ public class TripService {
     private final UserRepository userRepository;
 
     @Transactional
-    public void addTrip(TripRequestDto tripRequestDto) {
-        Region region = regionRepository.getOne(tripRequestDto.getRegionId());
-        User user = userRepository.getOne(tripRequestDto.getUserId());
+    public void addTrip(User user,TripRequestDto tripRequestDto) {
+        Region region = regionRepository.findById(tripRequestDto.getRegionId()).orElseThrow(() -> new CustomException(ErrorCode.DATA_NOT_FOUND));
+
         Trip trip = Trip.builder()
                 .region(region)
                 .user(user)
@@ -44,18 +43,26 @@ public class TripService {
     }
 
     @Transactional
-    public void deleteTrip(final long tripId) {
-        Trip ticket = tripRepository.getOne(tripId);
-        tripRepository.delete(ticket);
+    public int deleteTrip(User user, final long tripId) {
+        int resultCode = 200;
+        Trip ticket = tripRepository.findById(tripId).orElseThrow(() -> new CustomException(ErrorCode.DATA_NOT_FOUND));
+        if (!ticket.getUser().equals(user)) resultCode = 406;
+        else {
+            tripRepository.delete(ticket);
+        }
+        return resultCode;
     }
 
     @Transactional
-    public void updateTrip(final long tripId, TripRequestDto tripRequestDto) {
-        Trip trip = tripRepository.getOne(tripId);
-        Region region = regionRepository.getOne(tripRequestDto.getRegionId());
-        User user = userRepository.getOne(tripRequestDto.getUserId());
-
-        trip.update(tripRequestDto.getTitle(), tripRequestDto.getStartDate(), tripRequestDto.getEndDate(), user, region);
+    public int updateTrip(User user, final long tripId, TripRequestDto tripRequestDto) {
+        int resultCode = 200;
+        Trip trip = tripRepository.findById(tripId).orElseThrow(() -> new CustomException(ErrorCode.DATA_NOT_FOUND));
+        Region region = regionRepository.findById(tripRequestDto.getRegionId()).orElseThrow(() -> new CustomException(ErrorCode.DATA_NOT_FOUND));
+        if (!trip.getUser().equals(user)) resultCode = 406;
+        else {
+            trip.update(tripRequestDto.getTitle(), tripRequestDto.getStartDate(), tripRequestDto.getEndDate(), user, region);
+        }
+        return resultCode;
     }
 
     @Transactional
@@ -77,25 +84,26 @@ public class TripService {
     }
 
     @Transactional
-    public TripResponseDto findTrip(final User user,final long tripId){
-        Trip trip = tripRepository.getOne(tripId);
+    public TripResponseDto findTrip(final long tripId) {
+
+        Trip trip = tripRepository.findById(tripId).orElseThrow(() -> new CustomException(ErrorCode.DATA_NOT_FOUND));
         return new TripResponseDto(trip);
     }
 
     @Transactional
-    public TripResponseDto isInTrip(final User user, LocalDate todayDate){
+    public TripResponseDto isInTrip(final User user, LocalDate todayDate) {
         List<Trip> list = new ArrayList<>();
         list.addAll(tripRepository.findAllByUser(user));
-        for(Trip t : list){
-            if(todayDate.isAfter(t.getEndDate()))continue;
-            if(!todayDate.isBefore(t.getStartDate())&&todayDate.isBefore(t.getEndDate())){
+        for (Trip t : list) {
+            if (todayDate.isAfter(t.getEndDate())) continue;
+            if (!todayDate.isBefore(t.getStartDate()) && todayDate.isBefore(t.getEndDate())) {
                 return new TripResponseDto(t); //진행중일때
             }
         }
-        if(tripRepository.findFirstByStartDateAndUserOrderByStartDate(user,todayDate)==null){ //뒤 내용 아예없을때
+        if (tripRepository.findFirstByStartDateAndUserOrderByStartDate(user, todayDate) == null) { //뒤 내용 아예없을때
             return null;
-        }else{
-            return new TripResponseDto(tripRepository.findFirstByStartDateAndUserOrderByStartDate(user,todayDate).get(0));
+        } else {
+            return new TripResponseDto(tripRepository.findFirstByStartDateAndUserOrderByStartDate(user, todayDate).get(0));
         }
     }
 }
